requirements.txt
asgiref==3.8.1
cffi==1.16.0
cryptography==42.0.8
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
pycparser==2.22
pycryptodome==3.20.0
sqlparse==0.5.0
typing_extensions==4.12.2
